# backend/openapi.yaml
openapi: 3.1.0
info:
  title: WIT Sprint OS API
  version: "1.0.0"
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    REST API for the WIT Sprint OS delivery hierarchy.

    ## Naming

    The URL vocabulary predates the UI's, and the two do not match. Read it as:

    | URL / payload | The UI calls it |
    | ------------- | --------------- |
    | `product`     | **Module**      |
    | `module`      | **Component**   |

    The hierarchy is `client → project → product → module → sprint → task`,
    which the UI presents as Client → Project → Module → Component → Sprint.

    ## Conventions

    - **PATCH is partial.** Only the keys present in the body are written; an
      absent key leaves its column untouched. This is what makes two concurrent
      edits to different fields both survive. `PATCH /products/{id}/current-sprint`
      is the single exception where an explicit `null` differs from an absent
      key — see that operation.
    - **Ids may be supplied or generated.** Send `id` on create to choose it;
      omit it and the server mints one. A duplicate answers `409`.
    - **Lineage is enforced.** A child may only reference a parent inside its own
      product: a sprint's component, a task's backlog item, a report's sprint. A
      mismatch is a `400`, not a silent link.
    - **Timestamps** are RFC 3339 with an offset (`2026-07-23T02:35:08.59+07:00`).
...
      RFC 3339.
...

    ## Authentication

...
    it stands.

...
security: []
servers:
...
paths:
...
  /clients:
    get:
      tags: [clients]
      summary: List clients
      operationId: listClients
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: A page of clients.
          headers:
            X-Has-More:
              $ref: '#/components/headers/XHasMore'
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Client' }
        '503': { $ref: '#/components/responses/Unavailable' }
    post:
      tags: [clients]
      summary: Create a client
      operationId: createClient
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ClientCreate' }
      responses:
        '201':
          description: The created client.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Client' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /clients/{clientId}:
    parameters:
      - $ref: '#/components/parameters/ClientId'
    get:
      tags: [clients]
      summary: Get a client
      operationId: getClient
      responses:
        '200':
          description: The client.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Client' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [clients]
      summary: Update a client
      description: Partial — only the keys you send are written.
      operationId: updateClient
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ClientUpdate' }
      responses:
        '200':
          description: The updated client.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Client' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [clients]
      summary: Delete a client
      description: |
        Cascades to the client's projects, products, components, sprints,
        backlog, tasks, decisions and reports. There is no undo.
      operationId: deleteClient
      responses:
        '204': { description: Deleted. }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /clients/{clientId}/projects:
    parameters:
      - $ref: '#/components/parameters/ClientId'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Offset'
    get:
      tags: [clients, projects]
      summary: List a client's projects
      operationId: listProjectsByClient
      responses:
        '200':
          description: A page of projects.
          headers:
            X-Has-More: { $ref: '#/components/headers/XHasMore' }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Project' }
        '503': { $ref: '#/components/responses/Unavailable' }

  # ------------------------------------------------------------ projects ---
  /projects:
    get:
      tags: [projects]
      summary: List projects
      operationId: listProjects
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: A page of projects.
          headers:
            X-Has-More: { $ref: '#/components/headers/XHasMore' }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Project' }
        '503': { $ref: '#/components/responses/Unavailable' }
    post:
      tags: [projects]
      summary: Create a project
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectCreate' }
      responses:
        '201':
          description: The created project.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /projects/{projectId}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [projects]
      summary: Get a project
      operationId: getProject
      responses:
        '200':
          description: The project.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [projects]
      summary: Update a project
      description: |
        Moving a project to another client (`clientId`) carries its products
        with it — their denormalised `clientId` follows automatically.
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectUpdate' }
      responses:
        '200':
          description: The updated project.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [projects]
      summary: Delete a project
      description: Cascades to its products and everything beneath them.
      operationId: deleteProject
      responses:
        '204': { description: Deleted. }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/Unavailable' }

  # ---------------------------------------------- products (UI "Modules") ---
...
components:
  parameters:
    Limit:
      name: limit
      in: query
      description: Page size. Clamped to 1000; anything unparseable falls back to the default.
      schema: { type: integer, minimum: 1, maximum: 1000, default: 200 }
    Offset:
      name: offset
      in: query
      description: |
        Rows to skip. Postgres walks and discards them, so deep offsets get
        progressively more expensive — prefer narrowing by parent id. Values
        above 1,000,000 are refused with a 400 rather than silently answered
        with a different page.
      schema: { type: integer, minimum: 0, maximum: 1000000, default: 0 }
    ClientId:  { name: clientId,  in: path, required: true, schema: { type: string } }
    ProjectId: { name: projectId, in: path, required: true, schema: { type: string } }
...
  headers:
    XHasMore:
      description: Present and `true` when another page exists.
      schema: { type: string, enum: ['true'] }
  responses:
    BadRequest:
      description: The request is malformed, or a value is out of range.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: No such record.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: |
        A uniqueness clash, or a row was busy and the write should be retried.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    TooLarge:
      description: The body is over 1 MB.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unavailable:
      description: |
        Retryable. Either the request took too long — the statement timeout is
        5 s and the whole request is capped at 15 s — or the database could not
        take it: out of connection slots, restarting, or the connection dropped.
        Never a bug on the caller's side, and always worth retrying.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, examples: ['name is required'] }

    # ------------------------------------------------------------- enums ---
    ClientStatus:      { type: string, enum: [active, inactive, prospect] }
    Health:            { type: string, enum: [healthy, warning, at-risk] }
    Risk:              { type: string, enum: [low, medium, high] }
    ProjectStatus:     { type: string, enum: [discovery, active, done, on-hold] }
...
    Timestamp:
      type: string
      format: date-time
      description: RFC 3339 with an offset.
      examples: ['2026-07-23T02:35:08.590146+07:00']
...
    Json:
      type: [object, 'null']
      description: Free-form JSON, capped at 64 KB once decoded.

    # ----------------------------------------------------------- clients ---
    Client:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        industry: { type: string }
        status: { $ref: '#/components/schemas/ClientStatus' }
        clientPic: { type: string }
        witOwner: { type: string }
        contractType: { type: string }
        health: { $ref: '#/components/schemas/Health' }
        risk: { $ref: '#/components/schemas/Risk' }
        notes: { type: string }
        actionNeeded: { type: array, items: { type: string } }
        aiInsight: { $ref: '#/components/schemas/Json' }
        createdAt: { $ref: '#/components/schemas/Timestamp' }
        updatedAt: { $ref: '#/components/schemas/Timestamp' }
    ClientCreate:
      type: object
      required: [name]
      properties:
        id: { type: string, description: Omit to have one generated. }
        name: { type: string }
        industry: { type: string }
        status: { $ref: '#/components/schemas/ClientStatus' }
        clientPic: { type: string }
        witOwner: { type: string }
        contractType: { type: string }
        health: { $ref: '#/components/schemas/Health' }
        risk: { $ref: '#/components/schemas/Risk' }
        notes: { type: string }
        actionNeeded: { type: array, items: { type: string } }
        aiInsight: { $ref: '#/components/schemas/Json' }
    ClientUpdate:
      type: object
      description: Every field optional; absent keys are left alone.
      properties:
        name: { type: string }
        industry: { type: string }
        status: { $ref: '#/components/schemas/ClientStatus' }
        clientPic: { type: string }
        witOwner: { type: string }
        contractType: { type: string }
        health: { $ref: '#/components/schemas/Health' }
        risk: { $ref: '#/components/schemas/Risk' }
        notes: { type: string }
        actionNeeded: { type: array, items: { type: string } }
        aiInsight: { $ref: '#/components/schemas/Json' }

    # ---------------------------------------------------------- projects ---
    Project:
      type: object
      properties:
        id: { type: string }
        clientId: { type: string }
        name: { type: string }
        objective: { type: string }
        status: { $ref: '#/components/schemas/ProjectStatus' }
        createdAt: { $ref: '#/components/schemas/Timestamp' }
        updatedAt: { $ref: '#/components/schemas/Timestamp' }
    ProjectCreate:
      type: object
      required: [clientId, name]
      properties:
        id: { type: string }
        clientId: { type: string }
        name: { type: string }
        objective: { type: string }
        status: { $ref: '#/components/schemas/ProjectStatus' }
    ProjectUpdate:
      type: object
      properties:
        clientId: { type: string, description: 'Moves the project, and its products follow.' }
        name: { type: string }
        objective: { type: string }
        status: { $ref: '#/components/schemas/ProjectStatus' }

    # ------------------------------------------- products (UI "Modules") ---
...